% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_calendar.R
\name{Calendar}
\alias{Calendar}
\alias{d3_calendar}
\title{D3 Calendar Visual}
\usage{
d3_calendar(data, date, value, start, end, colors, showToday = TRUE,
  width = NULL, height = NULL)
}
\arguments{
\item{data}{A data.frame containing columns to use for date and value}

\item{date}{The column containing the dates to use for the calendar}

\item{value}{The column containing the values}

\item{start}{The start year for the calendar e.g. 2018}

\item{end}{The end year for the calendar e.g. 2020}

\item{colors}{A vector of colors to provide the interpolated color scale}

\item{showToday}{Mark out today's date?}

\item{width, height}{Paramters to override the default sizing policy}
}
\description{
Function to create a D3 calendar graphic.
}
