% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_dendrogram.R
\name{Dendrogram}
\alias{Dendrogram}
\alias{d3_dendrogram}
\title{D3 Dendrogram Visual}
\usage{
d3_dendrogram(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{Data for the dendrogram - TODO: give format}

\item{width, height}{Parameters to override the default sizing policy}
}
\description{
Function to create a D3 dendrogram graphic.
}
