% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3_radialtree.R
\name{RadialTree}
\alias{RadialTree}
\alias{d3_radialtree}
\title{D3 RadialTree Visual}
\usage{
d3_radialtree(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{Data for the radialtree - TODO: give format}

\item{width, height}{Parameters to override the default sizing policy}
}
\description{
Function to create a D3 radialtree graphic.
}
